+++ inscription/index/height.go
@@ -14,6 +14,6 @@ func (h *Height) N() uint32 {
 	return h.height
 }
 
-func (h *Height) Subsidy() uint32 {
+func (h *Height) Subsidy() uint64 {
 	return NewEpochFrom(h).Subsidy()
 }
